Prepare to release v1.0.0 (#64)
* Add v1.0.0 release notes
* Add public API contract
* Rev version number to v1.0.0
* Move logged_subprocess() from build to plugin_helers

---------
