+++ src/turnkeyml/common/build.py
@@ -18,7 +18,6 @@ import numpy as np
 import sklearn.base
 import turnkeyml.common.exceptions as exp
 import turnkeyml.common.tf_helpers as tf_helpers
-import turnkeyml.run.plugin_helpers as plugin_helpers
 from turnkeyml.version import __version__ as turnkey_version
 
 
@@ -443,65 +442,6 @@ class Logger:
     def flush(self):
         # needed for python 3 compatibility.
         pass
-
-
-def logged_subprocess(
-    cmd: List[str],
-    cwd: str = os.getcwd(),
-    env: Optional[Dict] = None,
-    log_file_path: Optional[str] = None,
-    log_to_std_streams: bool = True,
-    log_to_file: bool = True,
-) -> None:
-    """
-    This function calls a subprocess and sends the logs to either a file, stdout/stderr, or both.
-
-    cmd             Command that will run o a sbprocess
-    cwd             Working directory from where the subprocess should run
-    env             Evironment to be used by the subprocess (useful for passing env vars)
-    log_file_path   Where logs will be stored
-    log_to_file     Whether or not to store the subprocess's stdout/stderr into a file
-    log_to_std      Whether or not to print subprocess's stdout/stderr to the screen
-    """
-    if env is None:
-        env = os.environ.copy()
-    if log_to_file and log_file_path is None:
-        raise ValueError("log_file_path must be set when log_to_file is True")
-
-    log_stdout = ""
-    log_stderr = ""
-    try:
-        proc = subprocess.run(
-            cmd,
-            check=True,
-            env=env,
-            capture_output=True,
-            cwd=cwd,
-        )
-    except Exception as e:  # pylint: disable=broad-except
-        log_stdout = e.stdout.decode("utf-8")  # pylint: disable=no-member
-        log_stderr = e.stderr.decode("utf-8")  # pylint: disable=no-member
-        raise plugin_helpers.CondaError(
-            f"Exception {e} encountered, \n\nstdout was: "
-            f"\n{log_stdout}\n\n and stderr was: \n{log_stderr}"
-        )
-    else:
-        log_stdout = proc.stdout.decode("utf-8")
-        log_stderr = proc.stderr.decode("utf-8")
-    finally:
-        if log_to_std_streams:
-            # Print log to stdout
-            # This might be useful when this subprocess is being logged externally
-            print(log_stdout, file=sys.stdout)
-            print(log_stderr, file=sys.stdout)
-        if log_to_file:
-            log = f"{log_stdout}\n{log_stderr}"
-            with open(
-                log_file_path,
-                "w",
-                encoding="utf-8",
-            ) as f:
-                f.write(log)
 
 
 def get_system_info():
