+++ src/turnkeyml/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.4.0"
+__version__ = "1.0.0"
